+++ source/Relativity.Testing.Framework.Api.FunctionalTests/Strategies/Groups/GroupCreateStrategyFixture.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using FluentAssertions;
+﻿using FluentAssertions;
 using NUnit.Framework;
 using Relativity.Testing.Framework.Extensions;
 using Relativity.Testing.Framework.Models;
@@ -19,16 +18,9 @@ namespace Relativity.Testing.Framework.Api.FunctionalTests.Strategies
 		}
 
 		[Test]
-		public void Create_WithNull()
-		{
-			Assert.Throws<ArgumentNullException>(() =>
-				_sut.Create(null));
-		}
-
-		[Test]
 		public void Create_WithEmptyEntity()
 		{
-			var result = _sut.Create(new Group());
+			Group result = _sut.Create(new Group());
 
 			result.ArtifactID.Should().BePositive();
 			result.Name.Should().NotBeNullOrEmpty();
@@ -48,10 +40,18 @@ namespace Relativity.Testing.Framework.Api.FunctionalTests.Strategies
 				Client = client
 			};
 
-			var result = _sut.Create(entity.Copy());
+			Group result = _sut.Create(entity.Copy());
 
 			result.ArtifactID.Should().BePositive();
-			result.Should().BeEquivalentTo(entity, o => o.Excluding(x => x.ArtifactID).Excluding(x => x.Client.Number).Excluding(x => x.Client.Status.ArtifactID));
+			result.Should().BeEquivalentTo(
+				entity,
+				o => o.Excluding(group => group.ArtifactID)
+					.Excluding(group => group.Client.Number)
+					.Excluding(group => group.Client.Status.ArtifactID)
+					.Excluding(group => group.Type)
+					.Excluding(group => group.Actions)
+					.Excluding(group => group.Guids)
+					.Excluding(group => group.Meta));
 		}
 	}
 }