+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.6.0] - 2021-09-30
+
+### Added
+
+- Support for Group Manager V1 API methods: Create, Get, Update, Delete.
+
+### Changed
+
+- IGroupService.Update returns updated group.
+
 ## [5.5.0] - 2021-09-29
 
 ### Added